requirements.txt
addict==2.4.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
docopt==0.6.2
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-resources==5.12.0
jsonpatch==1.32
jsonpickle==3.0.1
jsonpointer==2.3
kiwisolver==1.4.4
matplotlib==3.7.1
mmcls==0.25.0
mmcv-full==1.7.1
munch==2.5.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
prettytable==3.7.0
py-cpuinfo==9.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
sacred==0.8.4
scipy==1.10.1
six==1.16.0
smmap==5.0.0
tomli==2.0.1
torch==1.10.1+cu113
torchvision==0.11.2
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
visdom==0.2.4
wcwidth==0.2.6
websocket-client==1.5.2
wrapt==1.15.0
yapf==0.33.0
zipp==3.15.0
